requirements.txt
Django==1.11.3
pytz==2017.2
gunicorn==19.7.1
psycopg2==2.7.1
whitenoise==3.3.1
dj-static==0.0.6
dj-database-url==0.4.2
reportlab==3.4.0
django-filter==1.1.0
django-widget-tweaks==1.4.2